stm32/i2c: add async, dual interrupt scaffolding.
